+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.9] - 2020-10-23
 ## Fixed
 - Member Profile Action icons remaining in hover state after hover is gone
 - Member Profile copy links copying object, not plain text